+++ packages/zosfiles/__tests__/__system__/methods/upload/__resources__/view_file_uss.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+account=$1
+host=$2
+port=$3
+user=$4
+password=$5
+ru=$6
+ussName=$7
+encoding=$8
+
+zowe zos-files view uf $ussName --encoding $encoding --host $host --port $port --user $user --password $password --ru $ru
+exit $?