+++ packages/zosjobs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the Zowe z/OS jobs SDK package will be documented in this file.
 
+## Recent Changes
+- Enhancement: Added "exec-started" and "exec-ended" to IJob return data from GetJobs.getJob [#2320](https://github.com/zowe/zowe-cli/pull/2320)
+
 ## `8.1.1`
 
 - BugFix: Updated peer dependencies to `^8.0.0`, dropping support for versions tagged `next`. [#2287](https://github.com/zowe/zowe-cli/pull/2287)